requirements.txt
certifi==2024.6.2
copilotkit==0.1.39
fastapi==0.115.11
langchain==0.3.20
langchain_community==0.3.19
langchain_core==0.3.42
langchain_openai==0.3.8
langgraph==0.3.5
matplotlib==3.10.1
mplfinance==0.12.10b0
numpy==2.2.3
Pillow==11.1.0
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.15.2
trustcall==0.0.34
typing_extensions==4.12.2
uvicorn==0.34.0
vaderSentiment==3.3.2
vaderSentiment==3.3.2
yfinance==0.2.40